2.0.0

+++ ice-server/src/main/java/com/ice/server/controller/IceAppController.java
@@ -33,7 +33,7 @@ public class IceAppController {
     }
 
     @RequestMapping(value = "/ice-server/app/edit", method = RequestMethod.POST)
-    public Long appEdit(@RequestBody IceApp app) {
+    public Integer appEdit(@RequestBody IceApp app) {
         if (app == null) {
             throw new ErrorCodeException(ErrorCode.INPUT_ERROR, "app");
         }
